requirements.txt
boto3==1.21.11
botocore==1.24.11
certifi==2021.10.8
charset-normalizer==2.0.12
crypto==1.4.1
decorator==5.1.1
idna==3.3
jmespath==0.10.0
Naked==0.1.31
numpy==1.22.2
pexpect==4.8.0
ptyprocess==0.7.0
PyMySQL==1.0.2
python-dateutil==2.8.2
pytimeparse==1.1.8
PyYAML==6.0
requests==2.27.1
s3transfer==0.5.2
shellescape==3.8.1
six==1.16.0
urllib3==1.26.8
validators==0.18.2
